requirements.txt
amqp==2.4.0
asn1crypto==0.24.0
backcall==0.1.0
backports.functools-lru-cache==1.5
bashplotlib==0.6.5
beautifulsoup4==4.6.0
billiard==3.5.0.5
bleach==3.0.2
bottle==0.12.16
cachetools==3.1.0
celery==4.2.1
certifi==2018.10.15
cffi==1.12.2
chardet==3.0.4
cheroot==6.5.2
CherryPy==18.1.0
Click==7.0
colour==0.1.5
conda==4.3.16
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
crank==0.8.1
cryptography==2.6.1
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2.1
django-allauth==0.38.0
django-appconf==1.0.2
-e git://github.com/franckbret/django-blog-quintet.git@3caa012227e3a80e262bda370c5c6ceee112d2ad#egg=django_blog_quintet
django-blog-zinnia==0.20
django-bower==5.2.0
django-braces==1.13.0
django-calendarium==1.3.4
django-ckeditor==5.6.1
django-classy-tags==0.8.0
django-cms==3.5.3
django-colorfield==0.1.15
django-contrib-comments==1.9.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-djconfig==0.8.0
django-easy-maps==0.9.3
django-el-pagination==3.1.0
django-eventtools==0.9.11
django-extensions==2.1.6
django-filer==1.3.2
django-filter==2.1.0
django-formtools==2.1
-e git+https://github.com/rodrigoamaral/django-fullcalendar.git@d116fcfc2093e57035932a9ea7c111d449fd7e4d#egg=django_fullcalendar
django-haystack==2.8.1
django-infinite-scroll-pagination==0.2.0
django-js-asset==1.1.0
django-libs==2.0.1
django-model-utils==3.1.2
django-modelcluster==4.3
django-mptt==0.8.7
django-polymorphic==2.0
django-private-chat==0.1.9
django-recurrence==1.6.0
django-registration==3.0
django-registration-redux==2.4
django-rest-auth==0.9.3
django-rest-swagger==2.0.7
django-scheduler==0.8.8
django-sekizai==0.10.0
django-spirit==0.7.0
django-tables2==2.0.5
django-tagging==0.4.6
django-taggit==0.22.2
django-treebeard==4.3
django-widget-tweaks==1.4.2
django-xmlrpc==0.1.8
djangocms-admin-style==1.2.9
djangocms-attributes-field==1.0.0
djangocms-column==1.9.0
djangocms-file==2.2.0
djangocms-googlemap==1.2.0
djangocms-installer==1.0.2
djangocms-link==2.3.1
djangocms-picture==2.1.3
djangocms-snippet==2.1.0
djangocms-style==2.1.0
djangocms-text-ckeditor==3.7.0
djangocms-video==2.1.1
djangopypi==0.4.4
djangorestframework==3.9.1
docker==3.7.2
docker-pycreds==0.4.0
docutils==0.14
draftjs-exporter==2.0.0
easy-thumbnails==2.5
emoji==0.5.1
environ==1.0
filebrowser-safe==0.5.0
filelock==3.0.10
flake8==3.5.0
Flask==1.0.2
future==0.17.0
fuzzywuzzy==0.17.0
geographiclib==1.49
geopy==1.18.1
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
grappelli-safe==0.5.1
gunicorn==19.9.0
howdoi==1.1.14
html5lib==0.999999999
httpie==1.0.2
httplib2==0.12.1
hupper==1.7
icalendar==4.0.3
idna==2.7
inflect==2.1.0
ipython==7.3.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
itypes==1.1.0
jaraco.functools==1.20
jaraco.itertools==3.0.0
jedi==0.13.3
Jinja2==2.10
kiwisolver==1.0.1
kombu==4.2.2.post1
lxml==4.2.5
Markdown==2.6.2
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
Mezzanine==4.3.1
# Editable install with no version control (mezzanine-events==0.1.4)
-e /home/mariusz1/Pobrane/mezzanine-events-0.1.4
mistune==0.7.4
more-itertools==4.3.0
mots-vides==2015.5.11
mutagen==1.42.0
numpy==1.15.0
oauthlib==2.1.0
olefile==0.45.1
openapi-codec==1.3.2
opencv-python==3.4.2.17
parso==0.3.4
PasteDeploy==2.0.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.2.0
pkginfo==1.5.0.1
plaster==1.0
plaster-pastedeploy==0.7
pluggy==0.8.1
portend==2.3
progressbar==2.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.3.0
pycosat==0.6.3
pycparser==2.19
pyflakes==1.6.0
pygame==1.9.6
Pygments==2.2.0
pyOpenSSL==19.0.0
pyparsing==2.3.0
pyquery==1.4.0
pyramid==1.10.4
pysqlite3==0.2.1
python-dateutil==2.7.3
python-google-calendar-api==0.5
python3-openid==3.1.0
pytz==2019.1
redis==3.2.1
regex==2018.11.22
repoze.lru==0.7
requests==2.20.0
requests-cache==0.4.13
requests-oauthlib==1.0.0
rsa==4.0
ruamel.yaml==0.15.89
scipy==1.1.0
selenium==3.141.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
stream-framework==1.4.0
tempora==1.14
tld==0.9.3
toml==0.10.0
tox==3.7.0
tqdm==4.24.0
traitlets==4.3.2
translationstring==1.3
tryton==5.2.0
TurboGears2==2.4.0
tzlocal==1.5.1
uni-slugify==0.1.4
unicode==2.7
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.24
venusian==1.2.0
vine==1.2.0
virtualenv==16.2.0
wagtail==2.0
wagtailmenus==2.8
wcwidth==0.1.7
webencodings==0.5.1
WebOb==1.8.5
websocket-client==0.56.0
websockets==3.2
Werkzeug==0.15.4
Whoosh==2.7.4
wikipedia==1.4.0
Willow==1.1
wolframalpha==3.0.1
xmltodict==0.11.0
zc.lockfile==1.4
zope.deprecation==4.4.0
zope.interface==4.6.0
